% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{base_query}
\alias{base_query}
\title{Base query to docusign}
\usage{
base_query(token = global_token(), base_url = get_base_url(), ...)
}
\arguments{
\item{token}{authorization token, a \code{httr2_token}, generated with \code{get_token()}}

\item{...}{additional paths to append via \code{req_url_path_append()}}
}
\value{
the JSON response from the query
}
\description{
Base query to docusign
}
