% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\docType{data}
\name{docusign_env}
\alias{docusign_env}
\title{Environment for storing auth token}
\format{
An object of class \code{environment} of length 0.
}
\usage{
docusign_env
}
\description{
Environment for storing auth token
}
\keyword{datasets}
