% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_account_id}
\alias{get_account_id}
\title{Get Account ID}
\usage{
get_account_id(token = global_token(), demo = Sys.getenv("docuSign_demo"))
}
\arguments{
\item{token}{authorization token, a \code{httr2_token}, generated with \code{get_token()}}

\item{demo}{use the developer demo sandbox?}
}
\value{
account id
}
\description{
Get Account ID
}
