% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_envelope_document}
\alias{get_envelope_document}
\title{Download a Document from an Envelope}
\usage{
get_envelope_document(
  account_id = NULL,
  envelope_id = NULL,
  document_id = NULL,
  filename = NULL,
  token = global_token(),
  base_url = get_base_url()
)
}
\arguments{
\item{account_id}{account id}

\item{envelope_id}{envelope id}

\item{document_id}{document id}

\item{filename}{file name/path to which the file should be written}

\item{token}{a \code{httr2_token} for authentication}

\item{base_url}{base URL for the query}
}
\value{
nothing, invisibly
}
\description{
Download a Document from an Envelope
}
