% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{get_token}
\alias{get_token}
\title{Create or refresh authorization token}
\usage{
get_token(demo = Sys.getenv("docuSign_demo"))
}
\arguments{
\item{demo}{use the developer demo sandbox?}
}
\value{
a \code{httr2_token} to be used for authentication
}
\description{
A copy of the token is stored in \code{.GlobalEnv} as \code{.docusign_token}
and is searched for when required
}
