% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{get_user}
\alias{get_user}
\title{Get a user's metadata}
\usage{
get_user(
  token = global_token(),
  demo = Sys.getenv("docuSign_demo"),
  overwrite_account_id = FALSE
)
}
\arguments{
\item{token}{authorization token, a \code{httr2_token}, generated with \code{get_token()}}

\item{demo}{use the developer demo sandbox?}

\item{overwrite_account_id}{store a new account id?}
}
\value{
a parsed JSON structure
}
\description{
Get a user's metadata
}
