% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{list_documents}
\alias{list_documents}
\title{List Envelope Documents}
\usage{
list_documents(
  account_id = NULL,
  envelope_id = NULL,
  token = global_token(),
  base_url = get_base_url()
)
}
\arguments{
\item{account_id}{account id}

\item{envelope_id}{envelope for which documents should be listed}

\item{token}{a \code{httr2_token} for authentication}

\item{base_url}{base URL for the query}
}
\value{
a parsed JSON structure
}
\description{
List Envelope Documents
}
