% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{list_envelopes}
\alias{list_envelopes}
\title{List envelopes}
\usage{
list_envelopes(
  account_id = NULL,
  from_date = "2020-01-01",
  token = global_token(),
  base_url = get_base_url()
)
}
\arguments{
\item{account_id}{account id}

\item{from_date}{earliest date to be searched}

\item{token}{a \code{httr2_token} for authentication}

\item{base_url}{base URL for the query

Use \code{get_user()} to identify the relevant \code{account_id}}
}
\value{
a parsed JSON structure
}
\description{
List envelopes
}
