% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{list_folders}
\alias{list_folders}
\title{List Folders}
\usage{
list_folders(
  account_id = NULL,
  token = global_token(),
  base_url = get_base_url()
)
}
\arguments{
\item{account_id}{account id}

\item{token}{a \code{httr2_token} for authentication}

\item{base_url}{base URL for the query}
}
\value{
a parsed JSON structure
}
\description{
List Folders
}
