% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{list_users}
\alias{list_users}
\title{List users}
\usage{
list_users(
  account_id = get_account_id(),
  token = global_token(),
  base_url = get_base_url()
)
}
\arguments{
\item{account_id}{account id}

\item{token}{a \code{httr2_token} for authentication}

\item{base_url}{base URL for the query

Use \code{get_user()} to identify the relevant \code{account_id}}
}
\value{
a parsed JSON structure
}
\description{
List users
}
