% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account.R
\name{store_account_id}
\alias{store_account_id}
\title{Store account id}
\usage{
store_account_id(aid, overwrite = FALSE)
}
\arguments{
\item{aid}{account id}

\item{overwrite}{overwrite an existing id?}
}
\value{
nothing
}
\description{
Store account id
}
