% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{store_token}
\alias{store_token}
\title{Store an authentication token as a hidden object}
\usage{
store_token(token = NULL)
}
\arguments{
\item{token}{\code{httr2_token} to be saved}
}
\value{
nothing, invisibly
}
\description{
Store an authentication token as a hidden object
}
