% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{template_query}
\alias{template_query}
\title{Templated Query}
\usage{
template_query(
  endpoint = "",
  account_id = get_account_id(),
  token = global_token(),
  base_url = get_base_url(),
  perform = TRUE
)
}
\arguments{
\item{endpoint}{endpoint to query, e.g. "envelopes" or a field URI}

\item{account_id}{account id}

\item{token}{a \code{httr2_token} for authentication}

\item{base_url}{base URL for the query}

\item{perform}{perform the query?}
}
\value{
a parsed JSON structure, or if \code{perform} is \code{FALSE}, the request
}
\description{
Templated Query
}
