% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_user}
\alias{get_user}
\title{Get a user's metadata}
\usage{
get_user(token = global_token(), base_url = "https://account-d.docusign.com/")
}
\arguments{
\item{token}{a \code{httr2_token} for authentication}

\item{base_url}{base URL for the query}
}
\value{
a parsed JSON structure
}
\description{
Get a user's metadata
}
